Switched to a cpu-side application of transformation
vertices are transformed in actionscript and uploaded as vertex buffers instead of being xformed on the gpu.  result is more actionscript load, but larger batches of sprites submitted at once (and hence fewer drawTriangles calls).

This approach means more potential for later optimization of static sprites. Jury is still out on whether this is a better approach overall, or some blended approach is appropriate.

Signed-off-by: Ely Greenfield <egreenfi@adobe.com>