Particles update:  Support for animated spritesheet particles, and more configurable particle generation
ParticleSymbols now acces assets as their source, rather than textures.  Assets can represent sprite sheets, so when a particle symbol points to a spritesheet asset, it will only render one cell of the sprite sheet at a time for each particle.

particles can now animate through a series of cells if the source asset is a sprite sheet. See the new animation properties on ParticleSymbol for more details.

Client code can now configure how a ParticleSymbol initializes new particles.  assign a function to the initializeParticleCallback property of ParticleSymbol -- it should be of the form:

function initParticle(data: ParticleData):void;

in this function, you can initialize x,y, direction (as an angle), speed (in pixels/millisecond), and angular velocity (in radians / millisecond).  For each particle generated.  Note that the data passed in gets reused, so don't assume any fields will be  initialized to defaults if you assign an init function.

Signed-off-by: Ely Greenfield <egreenfi@adobe.com>